+++ src/test/unit/specific_feature_tests/blocking_elements.js
@@ -121,7 +121,7 @@ test("table has br spacers via .rawHtml()", function () {
         children;
     wymeditor.rawHtml(tableHtml);
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
     children = $body.children();
 
     if (is_double_br_browser) {
@@ -148,8 +148,9 @@ test("table has br spacers via table insertion", function () {
     var wymeditor = jQuery.wymeditors(0),
         $body,
         children;
+
     wymeditor.rawHtml('');
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
     wymeditor.setCaretIn($body[0]);
     wymeditor.insertTable(2, 3, '', '');
 
@@ -181,7 +182,7 @@ test("p + table has br spacers via .rawHtml()", function () {
         children;
     wymeditor.rawHtml(pTableHtml);
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
     children = $body.children();
 
     if (is_double_br_browser) {
@@ -213,7 +214,7 @@ test("p + table has br spacers via table insertion", function () {
         children;
     wymeditor.rawHtml('<p>p1</p>');
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
 
     // Move the selector to the first paragraph
     first_p = $body.find('p')[0];
@@ -251,7 +252,7 @@ test("p + table + p has br spacers via .rawHtml()", function () {
         children;
     wymeditor.rawHtml(pTablePHtml);
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
     children = $body.children();
 
     expect(7);
@@ -274,7 +275,7 @@ test("p + table + p has br spacers via table insertion", function () {
         children;
     wymeditor.rawHtml('<p>p1</p><p>p2</p>');
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
 
     // Move the selector to the first paragraph
     first_p = $body.find('p')[0];
@@ -303,7 +304,7 @@ test("p + table + table + p has br spacers via .rawHtml()", function () {
         children;
     wymeditor.rawHtml(pTableTablePHtml);
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
     children = $body.children();
 
     expect(9);
@@ -328,7 +329,7 @@ test("p + table + table + p has br spacers via table insertion", function () {
         children;
     wymeditor.rawHtml('<p>p1</p><p>p2</p>');
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
 
     // Move the selector to the first paragraph
     first_p = $body.find('p')[0];
@@ -360,7 +361,7 @@ test("h1 + blockquote + pre has br spacers via .rawHtml()", function () {
         children;
     wymeditor.rawHtml(h1BlockquotePreHtml);
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
     children = $body.children();
 
     if (is_double_br_browser) {
@@ -397,7 +398,7 @@ test("br spacers aren't deleted when arrowing through them", function () {
         children;
     wymeditor.rawHtml(pTablePHtml);
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
 
     function checkLayout($body) {
         children = $body.children();
@@ -459,7 +460,7 @@ test("br spacers don't cause lots of blank p's when arrowing down", function () 
         children;
     wymeditor.rawHtml(pTableHtml);
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
 
     // Move the selector to the br before the table
     makeSelection(wymeditor, $body[0], $body[0], 1, 1);
@@ -496,7 +497,7 @@ test("br spacers don't cause lots of blank p's when arrowing up", function () {
         children;
     wymeditor.rawHtml(pTablePHtml);
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
 
     // Move the selector to the br after the table
     makeSelection(wymeditor, $body[0], $body[0], 3, 3);
@@ -525,7 +526,7 @@ test("br spacers stay in place when content is inserted- post-br", function () {
         children;
     wymeditor.rawHtml(tableHtml);
 
-    $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+    $body = wymeditor.$body();
 
     // Move the selector to the 2nd br (index 2)
     makeSelection(wymeditor, $body[0], $body[0], 2, 2);
@@ -559,7 +560,7 @@ if (!no_keypress_textnode_wrap_browser) {
             children;
         wymeditor.rawHtml(tableHtml);
 
-        $body = jQuery(wymeditor._doc).find('body.wym_iframe');
+        $body = wymeditor.$body();
 
         // Move the selector to the start
         moveSelector(wymeditor, $body[0]);